backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.0
numpy==1.26.4
faiss-cpu==1.13.2
pyarrow==15.0.0
